Renaming "dispatcher", thematically belonging to clonk
